+++ src/main/java/com/github/jlangch/venice/impl/VeniceInterpreter.java
@@ -1962,11 +1962,16 @@ public class VeniceInterpreter implements Serializable  {
 		return new VncFunction(name, params, macro) {
 			@Override
 			public VncVal apply(final VncList args) {
-				// TODO: currently disabled because it makes some problems with demo scripts
 				if (hasVariadicArgs()) {
 					if (args.size() < getFixedArgsCount()) {
 						try (WithCallStack cs = new WithCallStack(new CallFrame(name, params.getMeta()))) {
-							throw new ArityException(args.size(), getQualifiedName());
+							throw new ArityException(String.format(
+									"Wrong number of args (%d) passed to the variadic function %s that "
+										+ "requires at least %d args (%s).", 
+									args.size(), 
+									getQualifiedName(),
+									getFixedArgsCount(),
+									getParams().toString(true)));
 						}
 					}
 				}